move farmObject
